was seg faulting cuz the query result only got rebuilt on the event, which was put in queue that runs at the end of frame, causing dangling pointers, fixed that.

+++ src/World/experiments/calculus.hpp
@@ -11,8 +11,8 @@ struct calculus : System
     void Update(double dt) override;
 
   private:
-    void TwoDimensionalProjection(float increase);
-    void ThreeDimensionalProjection(float increase);
-    void FourDimensionalProjection(float increase);
+    void TwoDimensionalProjection(int increase);
+    void ThreeDimensionalProjection(int increase);
+    void FourDimensionalProjection(int increase);
 };
 } // namespace N
